requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone
langchain-pinecone
langchain_community
langchain_ollama
langchain_core
langchain-huggingface